Swapping to uint64 for frisbee message ID, and adding "FBEE" reserved bytes at start of message (instead of 0s)

+++ client.go
@@ -93,7 +93,7 @@ func (c *Client) StreamConnCh() <-chan *StreamConn {
 }
 
 // NewStreamConn creates a new StreamConn from the underlying frisbee.Conn
-func (c *Client) NewStreamConn(id uint32) *StreamConn {
+func (c *Client) NewStreamConn(id uint64) *StreamConn {
 	return c.conn.NewStreamConn(id)
 }
