+++ server_test.go
@@ -72,7 +72,7 @@ func TestServerRaw(t *testing.T) {
 		err := c.WriteMessage(&Message{
 			To:            16,
 			From:          32,
-			Id:            uint32(q),
+			Id:            uint64(q),
 			Operation:     protocol.MessagePing,
 			ContentLength: messageSize,
 		}, &data)
@@ -151,7 +151,7 @@ func BenchmarkThroughput(b *testing.B) {
 				err := frisbeeConn.WriteMessage(&Message{
 					To:            uint32(i),
 					From:          uint32(i),
-					Id:            uint32(q),
+					Id:            uint64(q),
 					Operation:     protocol.MessagePing,
 					ContentLength: messageSize,
 				}, &data)
@@ -219,7 +219,7 @@ func BenchmarkThroughputWithResponse(b *testing.B) {
 				err := frisbeeConn.WriteMessage(&Message{
 					To:            uint32(i),
 					From:          uint32(i),
-					Id:            uint32(q),
+					Id:            uint64(q),
 					Operation:     protocol.MessagePing,
 					ContentLength: messageSize,
 				}, &data)
